[GEOS-8392] WMTS service configured metdata will only appear in the capabilities document if INSPIRE extension is installed
