+++ src/extension/inspire/src/main/java/org/geoserver/inspire/wmts/WMTSExtendedCapabilitiesProvider.java
@@ -6,27 +6,14 @@ package org.geoserver.inspire.wmts;
 
 import org.geoserver.ExtendedCapabilitiesProvider;
 import org.geoserver.catalog.MetadataMap;
-import org.geoserver.catalog.WorkspaceInfo;
-import org.geoserver.config.ContactInfo;
 import org.geoserver.config.GeoServer;
 import org.geoserver.gwc.wmts.WMTSInfo;
 import org.geoserver.inspire.ViewServicesUtils;
-import org.geowebcache.GeoWebCacheException;
-import org.geowebcache.config.meta.ServiceContact;
-import org.geowebcache.config.meta.ServiceInformation;
-import org.geowebcache.config.meta.ServiceProvider;
-import org.geowebcache.conveyor.Conveyor;
 import org.geowebcache.io.XMLBuilder;
-import org.geowebcache.layer.TileLayer;
-import org.geowebcache.service.OWSException;
-import org.geowebcache.service.wmts.WMTSExtension;
-import org.geowebcache.storage.StorageBroker;
+import org.geowebcache.service.wmts.WMTSExtensionImpl;
 import org.xml.sax.Attributes;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.List;
 
 import static org.geoserver.inspire.InspireMetadata.CREATE_EXTENDED_CAPABILITIES;
 import static org.geoserver.inspire.InspireMetadata.LANGUAGE;
@@ -34,7 +21,7 @@ import static org.geoserver.inspire.InspireMetadata.SERVICE_METADATA_TYPE;
 import static org.geoserver.inspire.InspireMetadata.SERVICE_METADATA_URL;
 import static org.geoserver.inspire.InspireSchema.COMMON_NAMESPACE;
 
-public class WMTSExtendedCapabilitiesProvider implements WMTSExtension {
+public class WMTSExtendedCapabilitiesProvider extends WMTSExtensionImpl {
 
     public final static String VS_VS_OWS_NAMESPACE = "http://inspire.ec.europa.eu/schemas/inspire_vs_ows11/1.0";
     public final static String VS_VS_OWS_SCHEMA = "http://inspire.ec.europa.eu/schemas/inspire_vs_ows11/1.0/inspire_vs_ows_11.xsd";
@@ -113,61 +100,4 @@ public class WMTSExtendedCapabilitiesProvider implements WMTSExtension {
         String language = (String) serviceMetadata.get(LANGUAGE.key);
         ViewServicesUtils.addScenario1Elements(translator, metadataURL, mediaType, language);
     }
-
-    @Override
-    public ServiceInformation getServiceInformation() {
-        WMTSInfo gsInfo = geoserver.getService(WMTSInfo.class);
-        ServiceInformation gwcInfo = new ServiceInformation();
-        // add service information
-        gwcInfo.setTitle(gsInfo.getTitle());
-        gwcInfo.setTitle(gsInfo.getTitle());
-        gwcInfo.setDescription(gsInfo.getAbstract());
-        gwcInfo.getKeywords().addAll(gsInfo.keywordValues());
-        gwcInfo.setFees(gsInfo.getFees());
-        gwcInfo.setAccessConstraints(gsInfo.getAccessConstraints());
-        // add provider information
-        ServiceProvider serviceProvider = new ServiceProvider();
-        serviceProvider.setProviderName(gsInfo.getMaintainer());
-        serviceProvider.setProviderName(gsInfo.getOnlineResource());
-        // add contact information
-        ContactInfo gsContactInfo = geoserver.getSettings().getContact();
-        if (gsContactInfo != null) {
-            ServiceContact gwcContactInfo = new ServiceContact();
-            gwcContactInfo.setIndividualName(gsContactInfo.getContactPerson());
-            gwcContactInfo.setPositionName(gsContactInfo.getContactPosition());
-            gwcContactInfo.setAddressType(gsContactInfo.getAddressType());
-            gwcContactInfo.setAddressStreet(gsContactInfo.getAddress());
-            gwcContactInfo.setAddressCity(gsContactInfo.getAddressCity());
-            gwcContactInfo.setAddressPostalCode(gsContactInfo.getAddressPostalCode());
-            gwcContactInfo.setAddressCountry(gsContactInfo.getAddressCountry());
-            gwcContactInfo.setPhoneNumber(gsContactInfo.getContactVoice());
-            gwcContactInfo.setFaxNumber(gsContactInfo.getContactFacsimile());
-            gwcContactInfo.setAddressEmail(gsContactInfo.getAddressElectronicMailAddress());
-            serviceProvider.setServiceContact(gwcContactInfo);
-        }
-        gwcInfo.setServiceProvider(serviceProvider);
-        return gwcInfo;
-    }
-
-    @Override
-    public List<OperationMetadata> getExtraOperationsMetadata() throws IOException {
-        return null;
-    }
-
-    @Override
-    public Conveyor getConveyor(HttpServletRequest request, HttpServletResponse response, StorageBroker storageBroker) throws GeoWebCacheException, OWSException {
-        return null;
-    }
-
-    @Override
-    public boolean handleRequest(Conveyor conveyor) throws OWSException {
-        return false;
-    }
-
-    @Override
-    public void encodeLayer(XMLBuilder xmlBuilder, TileLayer tileLayer) throws IOException {
-
-    }
-
-
 }
