+++ src/gwc/src/main/resources/applicationContext.xml
@@ -6,6 +6,11 @@
  -->
 <!DOCTYPE beans PUBLIC "-//SPRING//DTD BEAN//EN" "http://www.springframework.org/dtd/spring-beans.dtd">
 <beans>
+
+  <bean id="wmtsCapabilitiesProvider" class="org.geoserver.gwc.wmts.WMTSCapabilitiesProvider">
+    <constructor-arg ref="geoServer"/>
+  </bean>
+
   <bean class="org.geoserver.platform.ModuleStatusImpl">
     <constructor-arg index="0" value="gs-gwc"/>
     <constructor-arg index="1" value="GeoWeb Cache"/>
